test: update bindings tests
