+++ test-e2e/src/lib.rs
@@ -1,7 +1,9 @@
 use mopro_ffi::{app, WtnsFn};
 
 rust_witness::witness!(multiplier2);
+rust_witness::witness!(multiplier2bls);
 rust_witness::witness!(keccak256256test);
+rust_witness::witness!(hashbenchbls);
 
 app!();
 
@@ -11,6 +13,8 @@ fn zkey_witness_map(name: &str) -> Result<WtnsFn, MoproError> {
     match name {
         "multiplier2_final.zkey" => Ok(multiplier2_witness),
         "keccak256_256_test_final.zkey" => Ok(keccak256256test_witness),
+        "hashbench_bls_final.zkey" => Ok(hashbenchbls_witness),
+        "multiplier2_bls_final.zkey" => Ok(multiplier2bls_witness),
         _ => Err(MoproError::CircomError("Unknown circuit name".to_string())),
     }
 }
