+++ test-e2e/tests/bindings/test_circom_multiplier2.kts
