+++ test-e2e/tests/test_generated_bindings.rs
@@ -1,13 +1,8 @@
 uniffi::build_foreign_language_testcases!(
-    "tests/bindings/test_mopro.swift",
-    "tests/bindings/test_mopro.kts",
-    // //    "tests/bindings/test_mopro.rb",
-    // //    "tests/bindings/test_mopro.py",
-    // "tests/bindings/test_mopro_keccak.swift",
-    // // "tests/bindings/test_mopro_keccak.kts", // FIXME: java.lang.OutOfMemoryError: Java heap space
-    // "tests/bindings/test_mopro_keccak_static.swift",
-    // "tests/bindings/test_mopro_keccak_static.kts",
-    // "tests/bindings/test_mopro_rsa.swift",
-    // // "tests/bindings/test_mopro_rsa.kts", // FIXME: java.lang.OutOfMemoryError: Java heap space
-    // // "tests/bindings/test_mopro_rsa_static.swift",
+    "tests/bindings/test_circom_multiplier2.swift",
+    "tests/bindings/test_circom_multiplier2.kts",
+    "tests/bindings/test_circom_keccak.swift",
+    "tests/bindings/test_circom_keccak.kts",
+    "tests/bindings/test_circom_multiplier2_bls.swift",
+    "tests/bindings/test_circom_multiplier2_bls.kts",
 );
